+++ include/wigeon/polyline2d.h
@@ -18,12 +18,12 @@ struct PolyLine2D {
   void push_back(const Point2D& point);
   void push_back(double x, double y);
 
-  int size() const;
+  size_t size() const;
 
   Rectangle2D bounding_box() const;
 
-  Point2D point(int i) const;
-  LineSegment2D edge(int i) const;
+  Point2D point(size_t i) const;
+  LineSegment2D edge(size_t i) const;
 
   std::vector<Point2D> points;
 };
